fix: 修复本仅删除文件夹后，再次调用会抛出异常的问题

+++ BKGalMgr/ViewModels/Pages/MigratePageViewModel.cs
@@ -52,7 +52,8 @@ public partial class MigratePageViewModel : ObservableObject
                 });
                 if (ret == false)
                 {
-                    Directory.Delete(gameNewPath, true);
+                    if (Directory.Exists(gameNewPath))
+                        Directory.Delete(gameNewPath, true);
                     return false;
                 }
